UniAdminTools

Documentation for UniAdminTools.

UniAdminTools.ProjAlloc.optimize_project_allocations

UniAdminTools.ProjAlloc.projalloc

CLI

UniAdminTools.ProjAlloc.projalloc — Function

projalloc --choices CHOICES
          --projects PROJECTS
          [--output "project_allocations.csv"]
          [--overall_objective "happiness - 0.5 * load"]
          [--rank_to_happiness "10 - 2^(ranking - 1) + 1"]
          [--assignments_to_load "num_assigned^2"]
          [--optimizer_time_limit 5]
          [--max_students_per_project 4]
          [--max_students_per_teacher 12]
          [--silent]

Intro

Compute optimal project allocations for student projects, using two csv files.

Options

--choices: A csv file with no header (data starting at the first row). The first column should be the student name, and the rest should be project choices (integer).

--projects: A csv file with no header (data starting at the first row). The first column should be the teacher name, and the second column should be the project name.

--output <"project_allocations.csv"::String>: The filename to save the output to.

--overall_objective <"happiness - 0.5 * load"::String>: A function that takes the total happiness and the total load and returns a single number. This will be maximized.

--rank_to_happiness <"10 - 2^(ranking - 1) + 1"::String>: Convert a student-assigned ranking into a happiness, which will be summed over students.

--assignments_to_load <"num_assigned^2"::String>: A function that takes the number of students assigned to each project and returns a number. This will be summed over projects.

--optimizer_time_limit <5::Int>: How long to spend optimizing the project allocations. Should usually find it pretty quickly (within 5 seconds), but you might try increasing this to see if it changes the results.

--max_students_per_project <4::Int>: The maximum number of students that can be assigned to a project.

--max_students_per_teacher <12::Int>: The maximum number of students that can be assigned to a teacher.

Flags

--silent: Don't print out information about the optimization process.

Internal

UniAdminTools.ProjAlloc.optimize_project_allocations — Function

optimize_project_allocations(data; kws...)

Find optimal project allocations using HiGHS, Ipopt, and Juniper.

choices: A filename (or data itself) where each row is one student's project choices, with the first column being the student name, and the next columns being their choices (second column being their first choice).

projects: A filename (or data itself) where each row is a project. The first column should be the project name, and the second column the teacher name. The row index of the project is its identifier in the choices table.

output_fname: The filename to save the output to. The default is project_allocations.csv.

overall_objective: A function that takes the total happiness and the total load and returns a single number. The default is to maximize happiness minus 0.5 times the load.

rank_to_happiness: Convert a student-assigned ranking into a happiness.

assignments_to_load: A function that takes the number of students assigned to each project and returns a number. The default is to return the square of the number of students.

optimizer_time_limit: How long to spend optimizing the project allocations. Should usually find it pretty quickly (within 5 seconds), but you might try increasing this to see if it changes the results.

max_students_per_project: The maximum number of students that can be assigned to a project.

max_students_per_teacher: The maximum number of students that can be assigned to a teacher.

verbose: Whether to print out information about the optimization process.